refactor(TRR/XTC): read sizes as checked unsigned values

+++ include/chemfiles/formats/XTC.hpp
@@ -27,9 +27,9 @@ class XTCFormat final : public Format {
 
   private:
     struct FrameHeader {
-        int32_t natoms; /* The total number of atoms */
-        int32_t step;   /* Current step number       */
-        float time;     /* Current time              */
+        size_t natoms; /* The total number of atoms */
+        size_t step;   /* Current step number       */
+        float time;    /* Current time              */
     };
 
     /// Read header of the Frame at the current position
